Make #transaction check nested tx depth before implicit COMMIT/ROLLBACK.
Previously, #transaction only checked to see that it was inside *any* tx
before ending (COMMIT/ROLLBACK) the open tx.  However, the ability of users
to explicitly end a tx inside the block means that the #transaction block
might end in an enclosing tx wherein the implicit COMMIT/ROLLBACK behavior
intended for the *nested* transaction is incorrect.

Example:

    dbh.transaction do |dbh|   # tx depth 1
      dbh.execute(STMT_1)
      dbh.transaction do |dbh| # tx depth 2
        dbh.execute(STMT_2)
        dbh.commit             # tx depth 1 (#rollback would do the same)
      end                      # tx depth 0 (implicit COMMIT -- to early)
      dbh.execute(STMT_3)
    end                        # implicit COMMIT of STMT_3 ignored!
                               # explicit would have been ignored, too