Make RDBI::Schema a deprecated alias for RDBI::Relation.
In SQL, a schema is a namespace for relational and non-relational objects.  In
RDBI, an RDBI::Schema is a representation of a relational object such as a
table, view or the "nonce relation" that is the projection of attributes
(columns) in a SQL query.

Move "schema" implementation to 'rdbi/relation.rb', and make 'rdbi/schema.rb'
issue a warning and alias Schema when loaded.